% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{norm_traces_LS}
\alias{norm_traces_LS}
\title{norm_traces_LS}
\usage{
norm_traces_LS(channelMeans, auto.detect = TRUE, channel1.baseline = TRUE,
  baseline_fluor = "Red", indicator_fluor = "Green")
}
\arguments{
\item{channelMeans}{a data frame containing average trace values for each fluorophore.}

\item{auto.detect}{logical. If TRUE, selects the fluorophore channel with the higher mean intensity as the baseline.}

\item{channel1.baseline}{logical. TRUE indicates that channel 1 contains the baseline fluorophore (default red).}

\item{baseline_fluor}{A character string to indicate the name of the baseline fluorophore.}

\item{indicator_fluor}{A character string to indicate the name of the indicator fluorophore.}
}
\value{
a data frame containing raw pixel intensity values for each fluorophore over time.
}
\description{
Reads a data frame of averaged (one-dimensional) pixel intensity values for two fluorophore channels (one as a baseline) from a linescan dataset and calculates normalized fluorescence intensity over time.
Returns a data frame with average intensity across time for each fluorophore, average intesity divided by the baseline (G/R), and normalized intensity (dG/R).
}
\seealso{
Other LineScan functions: \code{\link{load_2ptiff}},
  \code{\link{plot_2P_Ca}}, \code{\link{plot_2P_baseline}},
  \code{\link{plot_2P_dGR}}, \code{\link{plot_2P_norm}}
}
