% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_2P_baseline}
\alias{plot_2P_baseline}
\title{plot_2P_baseline}
\usage{
plot_2P_baseline(traces, baseline_fluor = "Red", indicator_fluor = "Green",
  ...)
}
\value{
a plot of the baseline fluorophore average raw pixel intensity values over time.
}
\description{
plot_2P_baseline
}
\seealso{
Other LineScan functions: \code{\link{load_2ptiff}},
  \code{\link{norm_traces_LS}}, \code{\link{plot_2P_Ca}},
  \code{\link{plot_2P_dGR}}, \code{\link{plot_2P_norm}}
}
