% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_2P_Ca}
\alias{plot_2P_Ca}
\title{plot_2P_Ca}
\usage{
plot_2P_Ca(traces, baseline_fluor = "Red", indicator_fluor = "Green", ...)
}
\value{
a plot of the calcium-sensitive fluorophore average raw pixel intensity values over time.
}
\description{
plot_2P_Ca
}
\seealso{
Other LineScan functions: \code{\link{load_2ptiff}},
  \code{\link{norm_traces_LS}},
  \code{\link{plot_2P_baseline}},
  \code{\link{plot_2P_dGR}}, \code{\link{plot_2P_norm}}
}
