% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_2P_norm}
\alias{plot_2P_norm}
\title{plot_2P_norm}
\usage{
plot_2P_norm(traces, baseline_fluor = "Red", indicator_fluor = "Green", ...)
}
\value{
a plot of the G/R normalized calcium-sensitive fluorophore average pixel intensity values over time.
}
\description{
plot_2P_norm
}
\seealso{
Other LineScan functions: \code{\link{load_2ptiff}},
  \code{\link{norm_traces_LS}}, \code{\link{plot_2P_Ca}},
  \code{\link{plot_2P_baseline}}, \code{\link{plot_2P_dGR}}
}
