% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{save_plot_2P}
\alias{save_plot_2P}
\title{save_plot_2P}
\usage{
save_plot_2P(filename = NULL, filetype = "png", width = 6, height = 4,
  units = "in", type = "cairo", res = 300)
}
\description{
save_plot_2P
}
