% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{trace_norm}
\alias{trace_norm}
\title{trace_norm}
\usage{
trace_norm(traces, auto.detect = TRUE, baseline_fluor = "red",
  indicator_fluor = "green")
}
\arguments{
\item{traces}{a data frame containing average trace values for each fluorophore.}

\item{auto.detect}{logical. If TRUE, selects the fluorophore channel with the higher mean intensity as the baseline.}

\item{baseline_fluor}{A character string to indicate the name of the baseline fluorophore.}

\item{indicator_fluor}{A character string to indicate the name of the indicator fluorophore.}
}
\value{
a data frame containing average trace values for each fluorophore and normalized intensity values.
}
\description{
Reads a data frame of averaged (one-dimensional) pixel intensity values for two fluorophore channels (one as a baseline)
and calculates normalized fluorescence intensity over time.
Returns a data frame with average intensity across time for each fluorophore, average intesity divided by the baseline (G/R), and normalized intensity (dG/R).
}
