% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{load_2ptiff}
\alias{load_2ptiff}
\title{load_2ptiff}
\usage{
load_2ptiff(tiff = "*.tif", ignore.source = TRUE, ignore.more = NULL)
}
\arguments{
\item{tiff}{Specify a name pattern to search (via Sys.glob) for TIFF files of interest within the same directory.}

\item{ignore.source}{logical. If TRUE, ignores TIFF files with "source" in the file name.}

\item{ignore.more}{A character string to ignore additional TIFF files with the given pattern in the file name.}
}
\value{
a data frame containing average trace values for each fluorophore
}
\description{
Reads raw pixel intensity values from TIFF files generated from two-photon linescans.
Converts two-dimensional measures of fluorescence intensity into one-dimensional intensity averages.
Returns a data frame with average traces across time for each fluorophore.
}
